+++ core/BusinessLogic/Response/ConfigInfo.cs
@@ -5,6 +5,7 @@ namespace DFM.BusinessLogic.Response
 	public class ConfigInfo
 	{
 		public Boolean? UseCategories { get; set; }
+		public Boolean? UseAccountsSigns { get; set; }
 		public Boolean? SendMoveEmail { get; set; }
 		public Boolean? MoveCheck { get; set; }
 		public Boolean? Wizard { get; set; }