+++ sql/7.4.0.0.sql
@@ -1,0 +1,2 @@
+alter table config
+	add UseAccountsSigns bit not null;